+++ src/Fable.Transforms/Global/Compiler.fs
@@ -48,10 +48,15 @@ type InlineExpr =
       Body: FSharpExpr
       FileName: string }
 
+type CompilerPlugins =
+    { CallPlugins: Map<Fable.EntityRef, System.Type>
+      MemberDeclarationPlugins: Map<Fable.EntityRef, System.Type> }
+
 type Compiler =
     abstract LibraryDir: string
     abstract CurrentFile: string
     abstract Options: CompilerOptions
+    abstract Plugins: CompilerPlugins
     abstract ImplementationFiles: IDictionary<string, FSharpImplementationFileContents>
     abstract GetRootModule: fileName: string -> string
     abstract GetEntity: Fable.EntityRef -> Fable.Entity
@@ -59,3 +64,28 @@ type Compiler =
     abstract AddWatchDependency: file: string -> unit
     abstract AddLog: msg:string * severity: Severity * ?range: SourceLocation
                         * ?fileName:string * ?tag: string -> unit
+
+[<AutoOpen>]
+module CompilerExt =
+    type Compiler with
+        member com.ApplyPlugin<'Plugin, 'Input>(plugins: Map<_,_>, atts: Fable.Attribute seq, input: 'Input, transform) =
+            if Map.isEmpty plugins then input
+            else
+                (input, atts) ||> Seq.fold (fun input att ->
+                    match Map.tryFind att.Entity plugins with
+                    | None -> input
+                    | Some plugin ->
+                        let plugin = System.Activator.CreateInstance(plugin, List.toArray att.ConstructorArgs) :?> 'Plugin
+                        let helper =
+                            { new PluginHelper with
+                                member _.LogWarning(msg, r) = com.AddLog(msg, Severity.Warning, ?range=r, fileName=com.CurrentFile)
+                                member _.LogError(msg, r) = com.AddLog(msg, Severity.Error, ?range=r, fileName=com.CurrentFile) }
+                        transform plugin helper input)
+
+        member com.ApplyMemberDeclarationPlugin(decl: Fable.MemberDecl) =
+            com.ApplyPlugin<MemberDeclarationPluginAttribute,_>
+                (com.Plugins.MemberDeclarationPlugins, decl.Info.Attributes, decl, fun p h i -> p.Transform(h, i))
+
+        member com.ApplyCallPlugin(atts, expr: Fable.Expr) =
+            com.ApplyPlugin<CallPluginAttribute,_>
+                (com.Plugins.CallPlugins, atts, expr, fun p h e -> p.Transform(h, e))