+++ src/Fable.Core/RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 4.0.0-nagareyama-alpha-001
+
+* Nagareyama
+
 ### 3.1.6
 
 * Fixed TypedArray.set signature @Shmew